+++ packages/@expo/cli/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### 🎉 New features
 
 - Add additional port check after build to ensure port is still available. ([#24315](https://github.com/expo/expo/pull/24315) by [@EvanBacon](https://github.com/EvanBacon))
+- Add support for bun as a package manager. ([#24344](https://github.com/expo/expo/pull/24344) by [@EvanBacon](https://github.com/EvanBacon))
 
 ### 🐛 Bug fixes